% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello to someone.}
\usage{
hello(name = "Robert")
}
\arguments{
\item{name}{A string of characters. (Default = "Robert")}
}
\value{
Prints a friendly greeting.
}
\description{
Say hello to someone.
}
\examples{
hello(name = "Andreas")
}
